platform/backend/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
python-multipart>=0.0.9
sqlalchemy>=2.0
asyncpg>=0.29
alembic>=1.13
psycopg2-binary>=2.9
redis>=5.0
pyyaml>=6.0
python-dotenv>=1.0
pydantic>=2.0
pydantic-settings>=2.0
